# perf: ~9% on schedule bench [pr]