* list comprehension into tuple for some busy lambdas
* some specific lru-caches


`python test/external/external_benchmark_schedule.py` best of 5, python 3.12 M3 Pro

branch
```
***** model tensor in     22.48 ms
***** model schedule in   64.39 ms
***** model opts(32) in   72.84 ms
***** model lower in      65.42 ms
***** model rewrite in   349.28 ms
***** model linearize in  88.81 ms
20089
all 664.33 ms
```

master
```
***** model tensor in     22.42 ms
***** model schedule in   64.44 ms
***** model opts(32) in   84.02 ms
***** model lower in      68.04 ms
***** model rewrite in   380.23 ms
***** model linearize in 108.50 ms
20089
all 728.78 ms
```